% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-rewrite.R
\name{make_rule}
\alias{make_rule}
\title{Make a AST transformation rule}
\usage{
make_rule(from, to)
}
\arguments{
\item{from}{A character string.}

\item{to}{A character string.}
}
\value{
A function that takes a language object and returns a language object.
}
\description{
Make a AST transformation rule
}
\examples{
library(rockstaR)

rule = make_rule("nothing", "NULL")
lyrics = "Rockstar is nothing"
rewrite(lyrics, list(rule))

}
