% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-rules.R
\name{rockstar-to-r-rules}
\alias{rockstar-to-r-rules}
\alias{default_rules}
\title{Mapping Rockstar operators into R operators}
\usage{
default_rules()
}
\description{
Mapping Rockstar operators into R operators
}
