% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-compile.R
\name{transpile}
\alias{transpile}
\title{Compile Rockstar code into R code}
\usage{
transpile(x, rules = default_rules())
}
\arguments{
\item{x}{A character string; the expression to transpile to R.}

\item{rules}{A list of rewriting rules. See [make_rule] for more detail.}
}
\value{
A character string.
}
\description{
Compile Rockstar code into R code
}
